fix(cgw): prevent invalid infra-group events and align groupsmap handling
- stop publishing infra-group Kafka events when venue->group lookup fails
- remove invalid sentinel pattern (`groupId = -1`) and use safe `uint32_t` group IDs in handlers/events
- include groupId in infra add/del debug logs for traceability
- keep groupsmap creation in storage layer and use DB-generated group IDs from PostgreSQL
- simplify ORM by removing groupsmap-specific `CreateRecord(..., out_id)` special-case from generic `orm.h`
- keep GroupsMapDB API typed for CGW-compatible 32-bit group IDs

+++ src/InfraGroupEvents.h
@@ -22,7 +22,7 @@ namespace OpenWifi {
 
 	inline void PrepareInfraGroupEventJson(Poco::JSON::Object &Payload,
 										   const std::string &EventType,
-										   std::uint64_t GroupId,
+										   std::uint32_t GroupId,
 										   const Types::StringVec &Infras = {}) {
 		Payload.set("type", EventType);
 		Payload.set("infra_group_id", std::to_string(GroupId));
@@ -38,7 +38,7 @@ namespace OpenWifi {
 	}
 
 	inline void PublishInfraGroupEvent(const std::string &EventType,
-									   std::uint64_t GroupId,
+									   std::uint32_t GroupId,
 									   const Types::StringVec &Infras = {}) {
 		Poco::JSON::Object Payload;
 		PrepareInfraGroupEventJson(Payload, EventType, GroupId, Infras);
